+++ camera/src/commonMain/kotlin/tech/kotlinlang/camera/CameraHelper.kt
@@ -2,6 +2,7 @@ package tech.kotlinlang.camera
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.ImageBitmap
 import tech.kotlinlang.camera.analyser.ImageAnalyser
 
 interface CameraHelper {
@@ -9,4 +10,6 @@ interface CameraHelper {
     fun <T> CameraContent(modifier: Modifier, imageAnalyser: ImageAnalyser<T>)
 
     fun enableFlash(switchOn: Boolean)
+
+    suspend fun captureImage(): ImageBitmap
 }